remove(csharp): remove dead C# post-parse recovery passes
- Drop `notnull` unwrapping, Unicode identifier span extension, scoped-lambda splitting/block recovery, and LINQ query-expression recovery.
- Grammar output now produces these structures natively, making post-parse rewrites permanently unreachable.
- Validated against a 1,700-file real-world corpus and 84-program LINQ battery; zero triggers observed.
- Consolidate shared `csharpReplaceNodeContents` helper, remove `csharpSimpleJoinQuerySpec` struct, and drop unused `bytes`/`unicode` imports.
- Remove associated unit tests and dead query recovery functions.
